fix(adk_deterministic_transfer): support deterministic transfer interrupt&resume
